Created a new fsl module.
