+++ osl/source_recon/__init__.py
@@ -1,4 +1,4 @@
 #!/usr/bin/python
 
 from .batch import *  # noqa: F401, F403
-from .rhino.utils import setup_fsl  # noqa: F401, F403
+from .rhino.fsl_wrappers import setup_fsl  # noqa: F401, F403
