+++ osl/source_recon/rhino/__init__.py
@@ -5,4 +5,5 @@
 """
 
 from .rhino import *  # noqa: F401, F403
+from .fsl_wrappers import *  # noqa: F401, F403
 from . import utils  # noqa: F401, F403
