+++ osl/source_recon/rhino/rhino.py
@@ -2349,66 +2349,3 @@ def recon_timeseries2niftii(
     )
 
     return out_nii_fname, reference_brain_fname
-
-
-def fsleyes(image_list):
-    """Displays list of niftii's using external command line call to fsleyes.
-
-    Parameters
-    ----------
-    image_list : string or tuple of strings
-        Niftii filenames or tuple of niftii filenames
-
-    Examples
-    --------
-    fsleyes(image)
-    fsleyes((image1, image2))
-    """
-
-    # check if image_list is a single file name
-    if isinstance(image_list, str):
-        image_list = (image_list,)
-
-    cmd = "fsleyes "
-
-    for img in image_list:
-        cmd += img
-        cmd += " "
-
-    cmd += "&"
-    print(cmd)
-    rhino_utils.system_call(cmd)
-
-
-def fsleyes_overlay(background_img, overlay_img):
-    """Displays overlay_img and background_img using external command line call
-    to fsleyes.
-
-    Parameters
-    ----------
-    background_img : string
-        Background niftii filename
-    overlay_img : string
-        Overlay niftii filename
-    """
-
-    if type(background_img) is str:
-        if background_img == "mni":
-            mni_resolution = int(nib.load(overlay_img).header.get_zooms()[0])
-            background_img = op.join(
-                os.environ["FSLDIR"],
-                "data/standard/MNI152_T1_{}mm_brain.nii.gz".format(mni_resolution),
-            )
-        elif background_img[0:3] == "mni":
-            mni_resolution = int(background_img[3])
-            background_img = op.join(
-                os.environ["FSLDIR"],
-                "data/standard/MNI152_T1_{}mm_brain.nii.gz".format(mni_resolution),
-            )
-
-    cmd = "fsleyes {} --volume 0 {} --alpha 100.0 --cmap red-yellow \
---negativeCmap blue-lightblue --useNegativeCmap &".format(
-        background_img, overlay_img
-    )
-
-    rhino_utils.system_call(cmd)
