+++ osl/source_recon/rhino/utils.py
@@ -64,26 +64,6 @@ def system_call(cmd, verbose=False):
     os.system(cmd)
 
 
-def setup_fsl(directory):
-    """Setup FSL.
-
-    Parameters
-    ----------
-    directory : str
-        Path to FSL installation.
-    """
-    if "FSLDIR" not in os.environ:
-        os.environ["FSLDIR"] = directory
-        os.environ["PATH"] = "{:s}/bin:{:s}".format(directory, os.getenv("PATH"))
-        os.environ["FSLOUTPUTTYPE"] = "NIFTI_GZ"
-
-
-def check_fsl():
-    """Check FSL is installed."""
-    if "FSLDIR" not in os.environ:
-        raise RuntimeError("Please setup FSL, e.g. with osl.source_recon.setup_fsl().")
-
-
 def get_gridstep(coords):
     """Get gridstep (i.e. spatial resolution of dipole grid) in mm.
